add InitDefaultPath
