+++ pkg/mosn/starter.go
@@ -48,7 +48,10 @@ type Mosn struct {
 // NewMosn
 // Create server from mosn config
 func NewMosn(c *config.MOSNConfig) *Mosn {
+	initializeDefaultPath(config.GetConfigPath())
 	initializePidFile(c.Pid)
+	initializeTracing(c.Tracing)
+	initializeMetrics(c.Metrics)
 
 	//get inherit fds
 	inheritListeners, reconfigure, err := server.GetInheritListeners()
@@ -61,9 +64,6 @@ func NewMosn(c *config.MOSNConfig) *Mosn {
 		// parse MOSNConfig again
 		c = config.Load(config.GetConfigPath())
 	}
-
-	initializeTracing(c.Tracing)
-	initializeMetrics(c.Metrics)
 
 	m := &Mosn{
 		config: c,
@@ -219,6 +219,9 @@ func (m *Mosn) Close() {
 	// close service
 	store.StopService()
 
+	// stop reconfigure domain socket
+	server.StopReconfigureHandler()
+
 	// stop mosn server
 	for _, srv := range m.servers {
 		srv.Close()
@@ -285,6 +288,10 @@ func initializeMetrics(config config.MetricsConfig) {
 
 func initializePidFile(pid string) {
 	server.SetPid(pid)
+}
+
+func initializeDefaultPath(path string) {
+	types.InitDefaultPath(path)
 }
 
 type clusterManagerFilter struct {
