+++ pkg/server/handler.go
@@ -773,59 +773,3 @@ func GetInheritListeners() ([]net.Listener, net.Conn, error) {
 
 	return listeners, uc, nil
 }
-
-func ReconfigureHandler() {
-	defer func() {
-		if r := recover(); r != nil {
-			log.DefaultLogger.Errorf("transferServer panic %v", r)
-		}
-		ReconfigureHandler()
-	}()
-	syscall.Unlink(types.ReconfigureDomainSocket)
-
-	l, err := net.Listen("unix", types.ReconfigureDomainSocket)
-	if err != nil {
-		log.StartLogger.Errorf("reconfigureHandler net listen error: %v", err)
-		return
-	}
-	defer l.Close()
-
-	log.DefaultLogger.Infof("reconfigureHandler start")
-
-	ul := l.(*net.UnixListener)
-	uc, err := ul.AcceptUnix()
-	if err != nil {
-		log.DefaultLogger.Errorf("reconfigureHandler Accept error :%v", err)
-		return
-	}
-	defer uc.Close()
-
-	log.DefaultLogger.Infof("reconfigureHandler Accept")
-
-	_, err = uc.Write([]byte{0})
-	if err != nil {
-		log.DefaultLogger.Errorf("reconfigureHandler %v", err)
-		return
-	}
-
-	reconfigure(false)
-}
-
-func isReconfigure() bool {
-	var unixConn net.Conn
-	var err error
-	unixConn, err = net.DialTimeout("unix", types.ReconfigureDomainSocket, 1*time.Second)
-	if err != nil {
-		log.DefaultLogger.Infof("not reconfigure: %v", err)
-		return false
-	}
-	defer unixConn.Close()
-
-	uc := unixConn.(*net.UnixConn)
-	buf := make([]byte, 1)
-	n, _ := uc.Read(buf)
-	if n != 1 {
-		return false
-	}
-	return true
-}
