+++ test/integrate/functiontest/faultinject_test.go
@@ -18,7 +18,7 @@ import (
 	_ "github.com/alipay/sofa-mosn/pkg/stream/sofarpc"
 	"github.com/alipay/sofa-mosn/test/integrate"
 	"github.com/alipay/sofa-mosn/test/util"
-	jsoniter "github.com/json-iterator/go"
+	"github.com/json-iterator/go"
 	"golang.org/x/net/http2"
 )
